refactor: move month display and name icon to AgendaTopBar
