+++ androidApp/src/main/java/com/example/tasky/android/agenda/screen/AgendaScreen.kt
@@ -1,51 +1,24 @@
 package com.example.tasky.android.agenda.screen
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowDropDown
-import androidx.compose.material3.DatePicker
-import androidx.compose.material3.DatePickerDialog
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme.typography
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.material3.rememberDatePickerState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
+import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.example.tasky.android.R
 import com.example.tasky.android.agenda.components.AgendaDayBar
+import com.example.tasky.android.agenda.components.AgendaTopBar
 import com.example.tasky.android.theme.Black
-import com.example.tasky.android.theme.Light
-import com.example.tasky.android.theme.LightBlue
 import com.example.tasky.android.theme.MyApplicationTheme
 import com.example.tasky.model.agenda.AgendaItem
 import kotlinx.collections.immutable.ImmutableList
@@ -61,7 +34,12 @@ import kotlinx.serialization.Serializable
 
 fun NavGraphBuilder.agendaScreen() {
     composable<Agenda> {
+        AgendaScreen(AgendaScreenState())
     }
+}
+
+fun NavController.navigateToAgenda() {
+    navigate(Agenda)
 }
 
 @Serializable
@@ -75,14 +53,11 @@ data class AgendaScreenState(
     val selectedDateOffset: Int = 0,
 )
 
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 private fun AgendaScreen(
     state: AgendaScreenState,
     modifier: Modifier = Modifier,
 ) {
-    var isDateDialogOpen by remember { mutableStateOf(false) }
-
     Column(
         modifier =
             modifier
@@ -90,49 +65,14 @@ private fun AgendaScreen(
                 .background(Black),
     ) {
         Spacer(modifier = Modifier.height(8.dp))
-        Row(
+        AgendaTopBar(
+            month = state.startDate.toLocalDateTime(TimeZone.currentSystemDefault()).month,
+            name = state.name,
             modifier =
                 Modifier
                     .fillMaxWidth()
                     .padding(horizontal = 20.dp),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                modifier =
-                    Modifier.clickable {
-                        isDateDialogOpen = true
-                    },
-            ) {
-                Text(
-                    state.startDate.toLocalDateTime(TimeZone.currentSystemDefault()).month.name,
-                    style = typography.bodyLarge,
-                    lineHeight = 19.2.sp,
-                    color = Color.White,
-                )
-                Icon(Icons.Filled.ArrowDropDown, contentDescription = null, tint = Color.White)
-            }
-
-            Box(
-                modifier =
-                    Modifier
-                        .size(36.dp)
-                        .background(Light, CircleShape),
-            ) {
-                Text(
-                    state.name,
-                    style = typography.bodyMedium,
-                    fontSize = 13.sp,
-                    lineHeight = 15.6.sp,
-                    color = LightBlue,
-                    modifier =
-                        Modifier.align(
-                            Alignment.Center,
-                        ),
-                )
-            }
-        }
+        )
         Spacer(modifier = Modifier.height(8.dp))
         Column(
             modifier =
@@ -141,8 +81,7 @@ private fun AgendaScreen(
                     .background(
                         Color.White,
                         RoundedCornerShape(topStart = 30.dp, topEnd = 30.dp),
-                    )
-                    .padding(top = 16.dp, start = 16.dp, end = 16.dp),
+                    ).padding(top = 16.dp, start = 16.dp, end = 16.dp),
         ) {
             AgendaDayBar(
                 (0..state.numberOfDateShown)
@@ -155,38 +94,6 @@ private fun AgendaScreen(
                 Modifier.fillMaxWidth(),
             )
             Spacer(modifier = Modifier.height(16.dp))
-        }
-    }
-
-    if (isDateDialogOpen) {
-        val datePickerState = rememberDatePickerState()
-        val confirmEnabled =
-            remember {
-                derivedStateOf { datePickerState.selectedDateMillis != null }
-            }
-
-        DatePickerDialog(
-            onDismissRequest = {
-                isDateDialogOpen = false
-            },
-            confirmButton = {
-                TextButton(
-                    onClick = {
-                        isDateDialogOpen = false
-                    },
-                    enabled = confirmEnabled.value,
-                ) {
-                    Text(stringResource(R.string.ok))
-                }
-            },
-            dismissButton = {
-                TextButton(onClick = { isDateDialogOpen = false }) { Text(stringResource(R.string.cancel)) }
-            },
-        ) {
-            DatePicker(
-                state = datePickerState,
-                modifier = Modifier.verticalScroll(rememberScrollState()),
-            )
         }
     }
 }
